refactor(app): standardized screens' name
