+++ app/auth/profile_picture_screen.tsx
