+++ app/welcome/final_screen.tsx
